fix: Support interactive commands and scripts

+++ internal/lefthook/runner/execute_unix.go
@@ -13,24 +13,11 @@ import (
 	"strings"
 
 	"github.com/creack/pty"
-	"github.com/mattn/go-isatty"
-
-	"github.com/evilmartians/lefthook/internal/log"
 )
 
 type CommandExecutor struct{}
 
 func (e CommandExecutor) Execute(opts ExecuteOptions) (*bytes.Buffer, error) {
-	stdin := os.Stdin
-	if opts.interactive && !isatty.IsTerminal(os.Stdin.Fd()) {
-		tty, err := os.Open("/dev/tty")
-		if err == nil {
-			defer tty.Close()
-			stdin = tty
-		} else {
-			log.Errorf("Couldn't enable TTY input: %s\n", err)
-		}
-	}
 	command := exec.Command("sh", "-c", strings.Join(opts.args, " "))
 	rootDir, _ := filepath.Abs(opts.root)
 	command.Dir = rootDir
@@ -42,18 +29,32 @@ func (e CommandExecutor) Execute(opts ExecuteOptions) (*bytes.Buffer, error) {
 
 	command.Env = append(os.Environ(), envList...)
 
-	p, err := pty.Start(command)
-	if err != nil {
-		return nil, err
+	var out *bytes.Buffer
+
+	if opts.interactive {
+		command.Stdout = os.Stdout
+		command.Stdin = os.Stdin
+		command.Stderr = os.Stderr
+		err := command.Start()
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		p, err := pty.Start(command)
+		if err != nil {
+			return nil, err
+		}
+
+		defer func() { _ = p.Close() }()
+		defer func() { _ = command.Process.Kill() }()
+
+		go func() { _, _ = io.Copy(p, os.Stdin) }()
+
+		out = bytes.NewBuffer(make([]byte, 0))
+		_, _ = io.Copy(out, p)
 	}
 
-	defer func() { _ = p.Close() }()
-	defer func() { _ = command.Process.Kill() }()
-
-	go func() { _, _ = io.Copy(p, stdin) }()
-
-	out := bytes.NewBuffer(make([]byte, 0))
-	_, _ = io.Copy(out, p)
+	defer command.Process.Kill()
 
 	return out, command.Wait()
 }
